Add support for MQTT broker credentials
Credentials are not stored in the database, but can be provided via
command line arguments or a JSON configuration file. By default, the
logger looks for ~/.credentials

Also change the HTTP log template to provide topic links that filter the
messages to show only the selected topic. This needs more UI support to
be useful.